[FEATURE] user terminate
- add user terminate
- add isEmployed field to user table
- change timeload field to type float

+++ timetracker.Server/API/User/Models/AddUserRequest.cs
@@ -6,7 +6,7 @@
         public string Surname { get; set; }
         public string Email { get; set; }
         public string Password { get; set; }
-        public int Timeload { get; set; }
+        public float Timeload { get; set; }
         public string Position { get; set; }
         public List<string> Permissions { get; set; }
     }
